[monorail] Hoist cached-build to root
It should have nothing to do with the static codegen flow now.
